修改：现在不需要上传vendor.js

+++ server/deploy/deploy-front.sh
@@ -17,13 +17,16 @@ npm run build
 
 # 由于vendor.js偏大（1M左右），服务器带宽却非常少，所以将其放在OSS上
 # TODO 目前仅有caibo.ren有证书
-cd ~
-sed -ibakvendorjs 's/\/static\/js\/vendor\.\w*\.js/https:\/\/s.charpty.com\/s\/vendor.js/' ${dist_dir}/index.html
-rm -rf ./bakvendorjs
-cp ${dist_dir}/static/js/vendor*.js ~/vendor.js
-cd ~
-python upload_s.py vendor.js
-
+if [[ $1 =~ "-uv" ]]
+then
+    cd ~
+    sed_re='s/\/static\/js\/vendor\.\w*\.js/https:\/\/s.charpty.com\/s\/vendor.js/'
+    sed -ibakvendorjs ${sed_re} ${dist_dir}/index.html
+    rm -rf ./bakvendorjs
+    cp ${dist_dir}/static/js/vendor*.js ~/vendor.js
+    cd ~
+    python upload_s.py vendor.js
+fi
 # 仅仅将html目录setfacl给site用户
 rm -rf ~/html-bak
 mkdir -p ~/html-bak
